录像使用时间用帧数 工坊保存录像详细信息

+++ GameAClient/Assets/App/Scripts/Model/Protocols/Record.cs
@@ -36,7 +36,7 @@ namespace GameA
         /// <summary>
         /// 
         /// </summary>
-        private float _usedTime;
+        private int _usedTime;
         /// <summary>
         /// 
         /// </summary>
@@ -166,7 +166,7 @@ namespace GameA
         /// <summary>
         /// 
         /// </summary>
-        public float UsedTime { 
+        public int UsedTime { 
             get { return _usedTime; }
             set { if (_usedTime != value) {
                 _usedTime = value;
